Cleaned up files, added SSR for Trending
