+++ components/Layout/Layout.tsx
@@ -1,6 +1,5 @@
 // Components
 import Header from "./Header";
-import Nav from "./Nav";
 import Footer from "./Footer";
 
 // Styles
