+++ components/Movie/Movie.module.css
@@ -1,181 +1,180 @@
 .container {
-    padding: 1rem;
-    margin: 1rem 0rem;
+  padding: 1rem;
+  margin: 1rem 0rem;
 
-    display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: flex-start;
-    align-items: flex-start;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  justify-content: flex-start;
+  align-items: flex-start;
 
-    border: 2px solid #cdcdcd;
-    border-radius: 10px
+  border: 2px solid var(--border-color);
+  border-radius: 10px;
 }
 
 .moviecontainer {
-    padding: 0rem;
-    margin: 0rem;
+  padding: 0rem;
+  margin: 0rem;
 
-    border-style: hidden;
-    border-color: black;
-    border-width: 2px;
-    border-radius: 10px;
+  border-style: hidden;
+  border-color: black;
+  border-width: 2px;
+  border-radius: 10px;
 
-    display: flex;
-    flex-direction: column;
-    flex-wrap: nowrap;
-    justify-content: flex-start;
-    align-items: flex-start;
-  
-    font-size: 1rem;
-    font-weight: 500;
-    text-align: left;
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  justify-content: flex-start;
+  align-items: flex-start;
+
+  font-size: 1rem;
+  font-weight: 500;
+  text-align: left;
 }
 
 .flexrow {
-    padding: 0rem;
-    margin: 0rem;
+  padding: 0rem;
+  margin: 0rem;
 
-    display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: flex-start;
-    align-items: center;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  justify-content: flex-start;
+  align-items: center;
 }
 
 .subflexrow {
-    padding: 0rem;
-    margin: 0rem;
+  padding: 0rem;
+  margin: 0rem;
 
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    align-items: stretch;
-
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: stretch;
 }
 
 .subflexcol {
-    padding: 0rem;
-    margin: 0rem 1rem 0rem 0rem;
+  padding: 0rem;
+  margin: 0rem 1rem 0rem 0rem;
 
-    display: flex;
-    flex-direction: column;
-    flex-wrap: nowrap;
-    justify-content: flex-start;
-    align-items: flex-start;
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  justify-content: flex-start;
+  align-items: flex-start;
 }
 
 .prodcontainer {
-    padding: 0.5rem 1rem;
-    margin: 0rem 1rem 0rem 0rem;
+  padding: 0.5rem 1rem;
+  margin: 0rem 1rem 0rem 0rem;
 
-    display: flex;
-    flex-direction: column;
-    flex-wrap: nowrap;
-    justify-content: space-evenly;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  justify-content: space-evenly;
+  align-items: center;
 
-    border: 2px solid #cdcdcd;
-    border-radius: 10px
+  border: 2px solid var(--border-color);
+  border-radius: 10px;
 }
 
 .movietitle {
-    padding: 0rem;
-    margin: 0rem;
+  padding: 0rem;
+  margin: 0rem;
 
-    flex-grow: 60;
-    flex-shrink: 1;
+  flex-grow: 60;
+  flex-shrink: 1;
 
-    font-size: 3rem;
-    font-weight: 500;
-    text-align: left;
-    color: #2571ff;
+  font-size: 3rem;
+  font-weight: 500;
+  text-align: left;
+  color: var(--highlight-color);
 }
 
 .movieid {
-    padding: 0.5rem 0.75rem;
-    margin: 0rem;
+  padding: 0.5rem 0.75rem;
+  margin: 0rem;
 
-    align-self: flex-end;
+  align-self: flex-end;
 
-    font-size: 0.75rem;
-    font-weight: light;
-    text-align: left;
-    color: #666666;
+  font-size: 0.75rem;
+  font-weight: light;
+  text-align: left;
+  color: var(--alt-text-color2);
 }
 
 .moviegenre {
-    padding: 0rem 0.5rem 0rem 0rem;
-    margin: 0rem;
+  padding: 0rem 0.5rem 0rem 0rem;
+  margin: 0rem;
 
-    font-size: 0.9rem;
-    font-weight: medium;
-    text-align: left;
-    color: #777777;
+  font-size: 0.9rem;
+  font-weight: medium;
+  text-align: left;
+  color: var(--alt-text-color);
 }
 
 .movierating {
-    padding: 0.1rem;
-    margin: 0rem;
+  padding: 0.1rem;
+  margin: 0rem;
 
-    display: flex;
-    align-items: flex-start;
+  display: flex;
+  align-items: flex-start;
 
-    font-size: 0.9rem;
-    font-weight: medium;
-    text-align: left;
-    color: #333333;
+  font-size: 0.9rem;
+  font-weight: medium;
+  text-align: left;
+  color: var(--alt-text-color);
 }
 
 .moviedesc {
-    padding: 0rem 5rem 0rem 0rem;
-    margin: 1.5rem 0rem;
+  padding: 0rem 5rem 0rem 0rem;
+  margin: 1.5rem 0rem;
 }
 
 .moviedata {
-    padding: 0.25rem 0rem;
-    margin: 0rem;
+  padding: 0.25rem 0rem;
+  margin: 0rem;
 
-    font-size: 0.9rem;
-    font-weight: medium;
-    text-align: left;
-    color: #777777;
+  font-size: 0.9rem;
+  font-weight: medium;
+  text-align: left;
+  color: var(--alt-text-color2);
 }
 
 .movieprod {
-    padding: 0rem;
-    margin: 0rem;
+  padding: 0rem;
+  margin: 0rem;
 
-    font-size: 1rem;
-    font-weight: 500;
-    text-align: center;
-    color: #333333;
+  font-size: 1rem;
+  font-weight: 500;
+  text-align: center;
+  color: var(--alt-text-color);
 }
 
 .linkbutton {
-    padding: 0.3rem;
-    margin: 0.5rem;
+  padding: 0.3rem;
+  margin: 0.5rem;
 
-    align-self: center;
+  align-self: center;
 
-    font-size: 1rem;
-    font-weight: medium;
-    text-align: center;
-    color: grey;
+  font-size: 1rem;
+  font-weight: medium;
+  text-align: center;
+  color: var(--alt-highlight-color);
 
-    border: 2px solid grey;
-    border-radius: 10px;
+  border: 2px solid var(--alt-highlight-color);
+  border-radius: 10px;
 
-    transition: color 0.5s, border-color 0.5s;
+  transition: color 0.5s, border-color 0.5s;
 }
 .linkbutton:hover,
 .linkbutton:focus {
-    color: #3571ff;
-    border-color: #2571ff;
+  color: var(--highlight-color);
+  border-color: var(--highlight-color);
 }
 
 .poster {
-    padding: 0rem;
-    margin: 0rem;
+  padding: 0rem;
+  margin: 0rem;
 }
