+++ components/Search/Search.tsx
@@ -2,19 +2,15 @@
 import React, { useState } from "react";
 
 // Functions
-import { Results } from "components/Results";
-import { Trending } from "components/Trending";
 import Searchbar from "./Searchbar";
 
 // Styles
 import sbstyles from "./Searchbar.module.css";
-import styles from "./Search.module.css";
 
 function Search() {
   const [showSearch, setShowSearch] = useState(false);
 
   const showSearchHandler = () => {
-    console.log("Open Search Bar");
     setShowSearch(true);
   };
