+++ components/Search/Searchbar.module.css
@@ -1,5 +1,5 @@
 .main {
-  padding: /* 1rem  */ 0rem;
+  padding: 0rem;
   margin: 0rem;
 }
