+++ components/Trending/Trending.tsx
@@ -1,5 +1,5 @@
 // Libraries
-import React, { useState, useEffect, useMemo } from "react";
+import React, { useState, useEffect } from "react";
 
 // Functions
 import { Results } from "components/Results";
@@ -8,35 +8,14 @@ import { Results } from "components/Results";
 import styles from "./Trending.module.css";
 
 // API data
-const api_key = "6c0d3427fcf17ddf80d04106c35a3a98";
-const url = `https://api.themoviedb.org/3/trending/movie/day?api_key=${api_key}`;
+/* const api_key = "6c0d3427fcf17ddf80d04106c35a3a98";
+const url = `https://api.themoviedb.org/3/trending/movie/day?api_key=${api_key}`; */
 
-function Trending() {
-  const [movies, setMovies] = useState([]);
+interface ITrendingProps {
+  movies: Array<Object>;
+}
 
-  const getTrending = async () => {
-    // const getTrending = async ( e ) => {
-
-    // e.preventDefault();
-
-    console.log("-- Loading Trending Movies --"); // Debugging purposes
-
-    try {
-      // Parse data and store into movies array
-      const res = await fetch(url);
-      const data = await res.json();
-      // console.log(data.results);
-      setMovies(data.results);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    // Runs function getTrending on page load. (Runs every time page is loaded)
-    getTrending();
-  }, []); // The null array here is required to prevent useEffect from spamming API database.
-
+function Trending({ movies }: ITrendingProps) {
   return (
     <main className={styles.main}>
       <div className={styles.titleBar}>
@@ -48,5 +27,4 @@ function Trending() {
   ); // End Return
 }
 
-// export default React.memo(Trending);
 export default React.memo(Trending);
