+++ pages/movie/index.tsx
@@ -1,6 +1,19 @@
 // Functions
 import { Trending } from "components/Trending";
 
-export default function Main() {
-  return <Trending />; // End return
+// API data
+const api_key = "6c0d3427fcf17ddf80d04106c35a3a98";
+const url = `https://api.themoviedb.org/3/trending/movie/day?api_key=${api_key}`;
+
+export default function Main({ data }) {
+  return <Trending movies={data.results} />; // End return
+}
+
+export async function getServerSideProps() {
+  const res = await fetch(url);
+  const data = await res.json();
+
+  return {
+    props: { data },
+  };
 }
