+++ styles/globals.css
@@ -4,7 +4,8 @@
   --border-color: #8899a6;
   --card-color: #282828;
   --text-color: #fafafa;
-  --alt-text-color: #aaaaaa;
+  --alt-text-color: #999999;
+  --alt-text-color2: #555555;
   --highlight-color: #2571ff;
   --alt-highlight-color: #888888;
   --fav-color: #ff5f5f;
